refactor: extract shared service config and dashboard components
- Add src/lib/services.ts as single source of truth for service metadata (colors, logos, field configs)
- Add src/components/ServiceAvatar.tsx — shared avatar used by Dashboard and Settings
- Extract Dashboard sub-components to src/components/dashboard/:
  - DonutChart, NextResetCard, ServiceDonutCard, SubscriptionPanel
- Remove SERVICE_COLORS/SERVICE_LOGOS duplication between pages
- Remove unused files: App.css, progress.tsx, *.ico, react.svg, tauri.svg, vite.svg
- Remove unused greet command from Rust backend
