--- src/App.css
@@ -1,1 +1,0 @@
-/* cleared - styles handled by Tailwind + shadcn/ui */
