--- src/assets/chatgpt.ico
@@ -1,1 +1,0 @@
-<!DOCTYPE html><html lang="en-US"><head><title>Just a moment...</title><meta http-equiv="Content-Type" content="text/html; charset=UTF-8"><meta http-equiv="X-UA-Compatible" content="IE=Edge"><meta name="robots" content="noindex,nofollow"><meta name="viewport" content="width=device-width,initial-scale=1"><style>*{box-sizing:border-box;margin:0;padding:0}html{line-height:1.15;-webkit-text-size-adjust:100%;color:#313131;font-family:system-ui,-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,"Helvetica Neue",Arial,"Noto Sans",sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol","Noto Color Emoji"}body{display:flex;flex-direction:column;height:100vh;min-height:100vh}.main-content{margin:8rem auto;padding-left:1.5rem;max-width:60rem}@media (width <= 720px){.main-content{margin-top:4rem}}#challenge-error-text{background-image:url("data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIHdpZHRoPSIzMiIgaGVpZ2h0PSIzMiIgZmlsbD0ibm9uZSI+PHBhdGggZmlsbD0iI0IyMEYwMyIgZD0iTTE2IDNhMTMgMTMgMCAxIDAgMTMgMTNBMTMuMDE1IDEzLjAxNSAwIDAgMCAxNiAzbTAgMjRhMTEgMTEgMCAxIDEgMTEtMTEgMTEuMDEgMTEuMDEgMCAwIDEtMTEgMTEiLz48cGF0aCBmaWxsPSIjQjIwRjAzIiBkPSJNMTcuMDM4IDE4LjYxNUgxNC44N0wxNC41NjMgOS41aDIuNzgzem0tMS4wODQgMS40MjdxLjY2IDAgMS4wNTcuMzg4LjQwNy4zODkuNDA3Ljk5NCAwIC41OTYtLjQwNy45ODQtLjM5Ny4zOS0xLjA1Ny4zODktLjY1IDAtMS4wNTYtLjM4OS0uMzk4LS4zODktLjM5OC0uOTg0IDAtLjU5Ny4zOTgtLjk4NS40MDYtLjM5NyAxLjA1Ni0uMzk3Ii8+PC9zdmc+");background-repeat:no-repeat;background-size:contain;padding-left:34px}</style><meta http-equiv="refresh" content="360"></head><body><div class="main-wrapper" role="main"><div class="main-content"><noscript><div class="h2"><span id="challenge-error-text">Enable JavaScript and cookies to continue</span></div></noscript></div></div><script>(function(){window._cf_chl_opt = {cvId: '3',cZone: 'openai.com',cType: 'managed',cRay: '9dbebca4ca21ebed',cH: '.h0fNNE9z07pD.WA1ybup7eRMrBuNc1BiwwDF5r81ys-1773443687-1.2.1.1-fwIiIaMNOaiHb4iJLS365H.mqnHYXj9TMhy1btB_LLVqZzIYKa4A6TjVccZzRPsF',cUPMDTk:"\/favicon.ico?__cf_chl_tk=02n0_kjgN3LaC6sBg07WiXyW2baiNKu_tda8GqTvLXM-1773443687-1.0.1.1-q_47dZQ0zCZhBB_CQfFhHLICKVE_4a1sL8TET0CN1fQ",cFPWv: 'g',cITimeS: '1773443687',cTplC:0,cTplO:0,cTplV:5,cTplB: '0',fa:"\/favicon.ico?__cf_chl_f_tk=02n0_kjgN3LaC6sBg07WiXyW2baiNKu_tda8GqTvLXM-1773443687-1.0.1.1-q_47dZQ0zCZhBB_CQfFhHLICKVE_4a1sL8TET0CN1fQ",md: 'tvdGhEuZA9V1.pPtlQN1HIwcnwwnKDNFFMJ2Lh9fr0I-1773443687-1.2.1.1-fmJ4X6wQVuGLxwrPKjrx_sXXgp3BQOqq90lVviUCjsU4cOFQDVfAuAJV6sUlT.vNDZZdLrZrfUAc.XVet..sBl9QnY6fgQ_UEfRYcC3E8wqDfPtxi4VgT8ZBxVI9rQAN.cVkXgZc3QUYfwt0b.UqN0SSJJFKR_m_JDzgWKEg3TFf5isU1NI_MCsnR7BRsHHGsZjFQbsbO8TrL3xK_2BHxvYN_fIU8OA9WzdeUj9mpNU6T2SgBZh52I_4iFh1Gik48CaLOBUObRv_XdZJpNgNH0D6p0KgZ8jkmZexuG_8mstYJV7d1P.XRaUQfmE5eO9hkO5utkTldcWMszWaevRT_DDIqpZ52jC24k000TQF.aBcrVI_0om5aexxoQu2tcyDsUzrZJFUlgqOCvgc22wX3DPg7WFfgUbFpAe1auUBlIdDtIzmYCQfxbtvJ1xGZpXcVXJWmXRTw0P3JXAmRCFrF6Qt3i4gHfeZsMe7zxVqy90tYSTunfWBD2dFS9lknW07YufBPFJeTuyUjBhyBtfFWINtk8HQtcfAMaTR6FrViPmD3fYI5mVqD_npAqyKrGrp2Cr_gADJbssgmHsSEC3c3Ktr8Nyfa1uMM56nUDUDhwRWJoEScL1prYI5Va4ocMBD22rKmLJ9DU7Cb4qcVPZYfYdk34VPf87GlHqasxLoooOWPqzJD5WSHwOSyn8ETDNzZTxuSo7Ef4BPfMviFMbYzVJqrI2l.IwXnvarepTvQ0WSkRDCQDm5dZ3D1pM8XyC.1oEj8tGQETN5YAUyMu0a0_pA8vy.dcOElT65n7fGXfWhFduu73pFTD3Bm0F7e063El4RKRGsYKt._5B.BlmWBh7BEoqYnuJmvEEHmJ.mDNZAtHqAuK0yOdpHK0g1bGT6E8YODBg195ij5CjxH.4LDJd09XeCvWHg7aY9NSLcbNjKMC1wni7EmQhtI91krBuRdBrV4zvNdGNmYRN7nvHeBg',mdrd: 'U8uTxeJqc10_wscQRVwNv1LVWJFF3ZHVpKo1lREui6A-1773443687-1.2.1.1-vjWKoavpm6QCw1KRjub1VwrwdJdIgOvnWqtO9nZOGnR60MieYaGFPoSK3iV6wxrD7httY3LCQL.n.c0j548q4lZsgTWvdXnYt4BqR9kCto1FeOnEQ7tXZ_fBskHZy1bUTomvgB6aM2Rg64irVw5GTnQJ5ywgt1Z1ELQkd0yuWgf5rpYqkRneJD2bNdFhwy8UkuWP9n3G2wUtM1m3x5vXFHlsJIw1rqfFLFL6Cs94_F.r5mznkJOkb.81yx0NJiB_jdUtdKLcwbMQTBODTU.OGGMYqeq63iWKsyldbz2oOylIwx30TNICC3lSPLfxPiYdH9lIZMrD3DHG4iXLVCl.2inyV4XVbckPPIrD_xEyvX_QzRp0l0mWusWqNoRwJQfwaAm3rjfhvptY.AAHBgRQCotppeqzhFXbrP2R1cRN9H1MdFQWtEF2clPuTfn3l74UCuVuWva1qtVM.4OqWTeonV6JiqTbYeBSc3bGEUz7aevD6rQsal.Re95pQ4bUJFIi0NEtWM6RwJMd.vIxCmgPLWcmJ3V6V9.I0Yqf73T69FyiFKY.Utn9OSehL6OiwOlCnUjfM6G8JzElyK8rnyDTZWd7HvfzbDrEeDm2Nkxk0un2E38L2CjF3D3QDMwNEQhuXlHFe6Xoa4hTikoqg3H5hnDZbo1F7CGxtLlKCIS6QHuo0VqCFHjzA7X3lOmcCKV8x39uepGdlGim49rSpT9XXycpsULH7n0Mi1.SumO856V1BVd3xypmQfKE3PsKBq8cOXQX4NPc8bRYQ8Uk2aiM9E_JPKveon2H0gWoQl3PpwLWPfhPhtmKsHckpTsQK7vEnvx.5eMxWEcO4fsu1MwN4gWaocEJQZI6LrEeosgPvDJebAxoKAh6yW4xZh2PXNn2z1anip9vMAiRiAlrOSjpHtM7JPzWtcIvZT9CgLYPQWxM4mQ6hJ0RTXt7M5BOwncz3WfoDe.iFnoaHzDxWyYdw4WP9v1TALAKouriDzE5YrU3Vv3afi5Ko1OaildKPeZiDuG2ymPhn16Hu9xLMCRD9gY4WjVpoM44lKfR7gVYlJ5rXatE6wdCeu8BE3FGXO6FXvU4OHg0T9_b4By_mpxpsI8Ocnc4Kwpxxc7_rnb6eWxZ6kHe99jVho84L.CKy9enZQCkuVKB6iutxJcKHRLA7jZktY1aDbh7IbWUhflqksIlDYFDtdhKv.8TI4YBVm.yO_1MFPkYoD.zpr5YvtPgwZLaKCPHQGPJl4wFK_AgHIPKLhcv6W4xt9diRxuDcCAJTsRm50eoTb7o2Lv1nvO.umwdpWPH5HVzH5TP5IqG6EInXVkIxcBYtTuldAr.Tj8_3ftIP6xdTA.o2o.Q7jIzwCVRicbMAYrgF2uZgnM2QDLvPjqp7jTQrv24ffbm.S5QVO18GbZ05v3xcBUYEFGeqQ5qOwY1oebiGKYoIth8sjwmWluZHS4XBN1QWqW_eQgfhmGmRr6iEp7gG_qYVhv6vcJdkTg779ZJkoOjasvqsHfPf8buYwUfRO2h2KSoqFqXFQYGOdAxJeespvLkx2gqB218A_h3UdOiOlrabYgytY8PpYesxpc3FffTlM1HB.F_ThDrglNdi9z05_2Xlx7GIQsh4hGGme1hbql_b2iQyUD6iZJ8uDZ9db5IiLiBDD7xrBAsL34MwIBmVyaK3DCWSf9968cYhs8.HrB2JE8FAvg57aSeKVvlUdPkdicYzx8hk70gwA4WOg7kdr3ci6AdVo3pjVbFNILJJfDAwWarukFvhS8u5f05zKlriw24ifin4k3FAyo5XCweUopPoyXOO5NVLRocloxFy64pmLNDs1ddJTVZmgJTtt9QgpY9oPurdpO7YRdUAlTqbTSLYLTxMdwWXrkeWDIxuSfSdcMEviV.t9S4a61eNqUDaLomzaoPVHh6lovF3tutvHN3d39zdK7BFSVrps.JSbzM22BKT5SuW3uwQCjIVS4u5fE2rjHFGsftXkNdeh6rJqwXgjwkvDbKc24KXaJS22iqq92gRNTY_QoIbhyqcpfhTSBzF6p4c6BDILbw5tzwVfn0uiz3FiAt6P68ShyN0WMUNqTEC2rUpSfzEZ8fWE5RMMS1zQTPG0fJBtpeb06Kk.oh.W95GOq3eOPlOPPQA.j_Krk4rTdISF6mFBYiSHSN6Uzc0MKsvfS_m7DM_8NKwy6c0tTS_hwduFMrD.tm2SL4pff8dXbROw2FtJw4vG2eyKAmSrZUtH25zHnVPbgZhG_Fe_KEEay_NEo8EiC0tl.RR6hoiH1b6nyPgnEnJSprQI4PSPGk_0p8XvKsif7WxdBo0fh0SrC_LwTnnM.AtKH7lG5g94BkXnbhy8ggyBq.7RbQDEhtk_39zYRxusP1DfwLcwyOaSNyzF37IniX9NQ6Gx7lD68Nz5854bhUqBSVOtzAz9a3mnq3njZ87l53XpAlB9zCybUQC80VRwGCT8i3P9SbwHHij4l.BwqSsKF4o9D57jr4CJUYtTIm_el.I2N9JtCtJw',};var a = document.createElement('script');a.src = '/cdn-cgi/challenge-platform/h/g/orchestrate/chl_page/v1?ray=9dbebca4ca21ebed';window._cf_chl_opt.cOgUHash = location.hash === '' && location.href.indexOf('#') !== -1 ? '#' : location.hash;window._cf_chl_opt.cOgUQuery = location.search === '' && location.href.slice(0, location.href.length - window._cf_chl_opt.cOgUHash.length).indexOf('?') !== -1 ? '?' : location.search;if (window.history && window.history.replaceState) {var ogU = location.pathname + window._cf_chl_opt.cOgUQuery + window._cf_chl_opt.cOgUHash;history.replaceState(null, null,"\/favicon.ico?__cf_chl_rt_tk=02n0_kjgN3LaC6sBg07WiXyW2baiNKu_tda8GqTvLXM-1773443687-1.0.1.1-q_47dZQ0zCZhBB_CQfFhHLICKVE_4a1sL8TET0CN1fQ"+ window._cf_chl_opt.cOgUHash);a.onload = function() {history.replaceState(null, null, ogU);}}document.getElementsByTagName('head')[0].appendChild(a);}());</script></body></html>
